Refacotred Box2D Organized the code in Box2D better and added some doxygen style commenting.
