+++ ENIGMAsystem/SHELL/Collision_Systems/Box2D/B2Dfunctions.cpp
@@ -14,20 +14,13 @@
 *** You should have received a copy of the GNU General Public License along
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
-#include "Universal_System/collisions_object.h"
-#include "Universal_System/instance_system.h" //iter
-//#include "Universal_System/roomsystem.h"
-#include "Collision_Systems/collision_mandatory.h" //iter
-#include "Universal_System/instance.h"
 
 #include <string>
 #include <stdio.h>
-#include <vector>
-using std::vector;
 //using namespace std;
 
 #include <Box2D/Box2D.h>
-#include "functions.h"
+#include "B2Dfunctions.h"
 bool systemPaused = false;
 
 static inline double r2d(double r) { return r * 180 / M_PI; }
@@ -36,8 +29,7 @@ static inline double r2d(double r) { return r * 180 / M_PI; }
 // 1) provide function overloads for interfacing with how studio is all fucked up so we don't have to limit ours
 // 2) box2d uses an inversed y-axis, thus why physics_fixture_get_angle() returns -radianstodegrees(angle)
 // 3) static objects when hit appear to me to move about one pixel at times, I can't tell if this is just my eyes playin tricks
-// 4) worlds must be manually updated untill someone writes an update loop for this system which updates the current
-//    rooms binded world, I suggest also leaving the option to continue allowing you to manually update your world
+// 4) leave the option to continue allowing you to manually update your world
 // 5) box2d's manual also states you should blend previous timesteps for updating the world with the current timestep
 //    in order to make the simulation run smoother
 // 6) box2d manual is available here... http://www.box2d.org/manual.html
@@ -45,32 +37,8 @@ static inline double r2d(double r) { return r * 180 / M_PI; }
 // 8) box2d's classes allow you to set a b2Shape for instance to a b2CircleShape or b2PolygonShape
 //    that is why I wrote the classes to use an abstracted pointer reference such as b2Shape and b2Joint
 
-struct worldInstance {
-  b2World* world;
-  // Prepare for simulation. Typically we use a time step of 1/60 of a
-  // second (60Hz) and 10 iterations. This provides a high quality simulation
-  // in most game scenarios.
-  float32 timeStep;
-  int32 velocityIterations;
-  int32 positionIterations;
-  int32 pixelstometers;
-  bool paused;
-
-  worldInstance()
-  {
-    // Define the gravity vector.
-    b2Vec2 gravity(0.0f, 10.0f);
-
-    // Construct a world object, which will hold and simulate the rigid bodies.
-    world = new b2World(gravity);
-    timeStep = 1.0f / 60.0f;
-    velocityIterations = 8;
-    pixelstometers = 32;
-    paused = false;
-  }
-
-  void world_update();
-}; 
+vector<worldInstance> worlds(1);
+vector<fixtureInstance> fixtures;
 
 void worldInstance::world_update() 
 {
@@ -79,70 +47,6 @@ void worldInstance::world_update()
     world->ClearForces();
   }
 }
-
-vector<worldInstance> worlds(1);
-
-struct fixtureInstance {
-  int world;
-  b2Body* body;
-  b2Fixture* fixture;
-  b2Shape* shape;
-  b2PolygonShape* polygonshape;
-  vector<b2Vec2> vertices;
-
-  fixtureInstance() 
-  {
-
-  }
-
-  ~fixtureInstance()
-  {
-
-  }
-
-  void FinishShape()
-  {
-    b2FixtureDef fixtureDef;
-    fixtureDef.shape = shape;
-    fixture = body->CreateFixture(&fixtureDef);
-  }
-
-}; 
-vector<fixtureInstance> fixtures;
-
-struct jointInstance {
-  int world;
-  b2Joint* joint;
-  jointInstance()
-  {
-  }
-}; 
-vector<jointInstance> joints;
-
-/* This is just place holder shit to get the fuckin linker shut the fuckin hell up */
-enigma::instance_t collision_rectangle(double x1, double y1, double x2, double y2, int obj, bool prec, bool notme)
-{
-
-}
-
-namespace enigma
-{
-  void *get_collision_mask(sprite* spr, unsigned char* input_data, collision_type ct) // It is called for every subimage of every sprite loaded.
-  {
-    return 0;
-  }
-
-  void free_collision_mask(void* mask)
-  {
-  }
-
-  void perform_automatic_collision_handling_before_collision_event()
-  {
-    for (std::vector<worldInstance>::iterator it = worlds.begin(); it != worlds.end(); it++) {
-      it->world_update();
-    }
-  }
-};
 
 bool place_free(double x, double y)
 {
@@ -160,20 +64,18 @@ bool place_meeting(double x, double y, int object)
   return false;
 }
 
-/* This is where the actual physics_* functions start at */
-
 void physics_world_create(int pixeltometerscale)
 {
-  // studio's fucked up world creation just auto binds it to the current room
-  // thats fuckin retarded thus why i overloaded the function and provided an extra
-  // function for setting the pixel to metre scale, because i refuse to write system their way
+  /** studio's fucked up world creation just auto binds it to the current room
+      thats fuckin retarded thus why i overloaded the function and provided an extra
+      function for setting the pixel to metre scale, because i refuse to write system their way
+  **/
 }
 
 int physics_world_create()
 {
   int i = worlds.size();
-  worlds.resize(i + 1);
-  worlds[i] = worldInstance();
+  worlds.push_back(worldInstance());
   return i;
 }
 
@@ -285,12 +187,11 @@ int physics_fixture_create(int world)
   else
   {
     int i = fixtures.size();
-    fixtures.resize(i + 1);
     fixtureInstance fixture = fixtureInstance();
     b2BodyDef bodyDef;
     bodyDef.type = b2_dynamicBody;
     fixture.body = worlds[world].world->CreateBody(&bodyDef);
-    fixtures[i] = fixture;
+    fixtures.push_back(fixture);
     fixtures[i].world = world;
     return i;
   }
@@ -755,65 +656,6 @@ void physics_apply_local_torque(int id, double torque)
     fixtures[id].body->ApplyTorque(torque);
   }
 }
-
-int physics_joint_create(int world)
-{
-    int i = joints.size();
-    joints.resize(i + 1);
-    jointInstance joint = jointInstance();
-    joint.world = world;
-    joints[i] = joint;
-    return i;
-}
-
-void physics_joint_distance_create(int id, int fixture1, int fixture2)
-{
-  if (unsigned(id) >= joints.size() || id < 0)
-  {
-    return;
-  }
-  else
-  {
-    b2JointDef jointDef;
-    jointDef.bodyA = fixtures[fixture1].body;
-    jointDef.bodyB = fixtures[fixture2].body;
-    //jointDef.target.Set(350, 320);
-    //jointDef.maxForce = 3000.0 * body.m_mass;
-    //jointDef.timeStep = m_timeStep;
-    joints[id].joint = worlds[joints[id].world].world->CreateJoint(&jointDef);
-  }
-}
-
-void physics_joint_mouse_create(int id, int fixture)
-{
-  if (unsigned(id) >= joints.size() || id < 0)
-  {
-    return;
-  }
-  else
-  {
-    b2MouseJointDef jointDef;
-    //jointDef.body1 = m_world.m_groundBody;
-    //jointDef.body2 = body;
-    //jointDef.target.Set(350, 320);
-    //jointDef.maxForce = 3000.0 * body.m_mass;
-    //jointDef.timeStep = m_timeStep;
-    joints[id].joint = worlds[joints[id].world].world->CreateJoint(&jointDef);
-  }
-}
-
-void physics_joint_set_target(int id, double x, double y)
-{
-  if (unsigned(id) >= joints.size() || id < 0)
-  {
-    return;
-  }
-  else
-  {
-    //joints[id].joint->SetTarget(b2Vec2(x, y));
-  }
-}
-
 
 void physics_pause_enable(bool pause)
 {
