+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3d3d.h
@@ -51,12 +51,31 @@ extern bool d3dMode;
 extern double projection_matrix[16];
 extern double transformation_matrix[16];
 
+/** Starts 3D mode.
+**/
 void d3d_start();
+/** Ends 3d mode.
+**/
 void d3d_end();
+/** Gets whether 3D mode is enabled.
+@return This function returns true if 3D mode is enabled, and false if it is not.
+**/
 bool d3d_get_mode();
+/** Sets the use of a perspective projection.
+@param enable whether or not to enable perspective
+**/
 void d3d_set_perspective(bool enable);
+/** Sets whether hidden surface removal is enabled.
+@param enable whether or not to enable hidden surface removal
+**/
 void d3d_set_hidden(bool enable);
+/** Sets whether depth buffer writing is enabled.
+@param enable whether or not to enable depth writing
+**/
 void d3d_set_zwriteenable(bool enable);
+/** Sets whether lighting is enabled.
+@param enable whether or not to enable lighting
+**/
 void d3d_set_lighting(bool enable);
 
 void d3d_set_culling(bool enable);
